app/vlselect/internalselect: explicitly verify for the missing query arg at getInt64FromRequest() function
This function expects that the given query arg must be always set by in the incoming request.
If it isn't set, then the corresponding error must be returned, so users could quickly identify and fix it.
Previously the error was slightly misleading for this case: 'cannot parse the arg value as integer: ...'